✨ (manga): Migrate search functionality to a new module
This commit migrates the search functionality from the old `search` module to a new `search-manga` module. The new module uses the Elasticsearch service to index and search manga data.  It also introduces an indexing service that indexes all manga data on application startup. This improves the search performance and maintainability of the code. The suggest functionality has been temporarily removed and will be re-implemented in a future update.