🚀 (deps): migrate from Azure Blob Storage to AWS S3
This commit migrates the image storage from Azure Blob Storage to AWS S3.  This change was made due to cost considerations and the better integration of AWS S3 with other services we are using.  The `AzureService` and `BlobModule` are replaced with `S3Service` and `S3Module`, respectively.  The cron job schedule is also adjusted to run at 23:15 Bangkok time.